+++ vispy/scene/transforms/__init__.py
@@ -1,9 +1,17 @@
 # -*- coding: utf-8 -*-
 # Copyright (c) 2014, Vispy Development Team.
 # Distributed under the (new) BSD License. See LICENSE.txt for more info.
+"""
+Provides classes representing different transform types suitable for
+use with visuals and scenes.
+"""
+
+__all__ = ['NullTransform', 'STTransform', 'AffineTransform',
+           'PerspectiveTransform', 'LogTransform', 'PolarTransform',
+           'InvPolarTransform', 'ChainTransform']
 
 from .base_transform import BaseTransform  # noqa
-from .linear import (NullTransform, STTransform, SRTTransform,  # noqa
+from .linear import (NullTransform, STTransform,  # noqa
                      AffineTransform,  PerspectiveTransform)  # noqa
 from .nonlinear import LogTransform, PolarTransform, InvPolarTransform  # noqa
 from .chain import ChainTransform  # noqa